backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-docx>=1.1.0
PyMuPDF>=1.23.0
latex2mathml>=3.0.0
mistune>=3.0.0
Jinja2>=3.1.2
openai>=1.6.0
sqlalchemy[asyncio]>=2.0.23
aiosqlite>=0.19.0
python-multipart>=0.0.6
sse-starlette>=1.8.0
pydantic-settings>=2.1.0
pytest>=7.0
pytest-asyncio>=0.23.0
